requirements.txt
streamlit
requests
Pillow
torch
transformers
numpy
openai
chromadb>=0.4.15
tiktoken